add BaseAssistantClient
